add formula

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        double fahrenheit = ((temperatureCelsius * 9) / 5 ) + 32;
+        double fahrenheit =  (((double)temperatureCelsius * 9) / 5 ) + 32;
         System.out.println(fahrenheit);
     }
 }
